modified:   calculate-heart-rate/patch.cpp 	modified:   calculate-heart-rate/patch.xodp 	new file:   example-test-heart-rate-monitor/patch.xodp 	modified:   heart-rate-monitor/patch.xodp 	deleted:    test-state-patch/patch.xodp 	deleted:    test-state/patch.cpp 	deleted:    test-state/patch.xodp

+++ calculate-heart-rate/patch.cpp
@@ -18,7 +18,7 @@ void evaluate(Context ctx) {
     // get state
     State* state = getState(ctx);
 
-    int32_t irValue = getValue<input_IR>(ctx);
+    long irValue = getValue<input_IR>(ctx);
     
     if (checkForBeat(irValue) == true)
     {
